requirements.txt
accelerate==0.15.0
numpy==1.26.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.4
tqdm==4.66.1
Pillow==10.1.0
PyYAML==6.0.1
pandas==2.1.4
requests==2.31.0
urllib3==2.1.0
pytorch-metric-learning==2.4.1
kaggle==1.6.6
scikit-learn==1.3.2
scipy==1.11.4
yacs==0.1.8
timm==0.4.12

